+++ src/app/components/venue-details-card/venue-details-card.component.ts
@@ -2,13 +2,16 @@ import { Component,
          EventEmitter, 
          Input, 
          OnInit, 
-         Output }              from '@angular/core';
+         Output }                    from '@angular/core';
 
-import { Venue, VenueDetails } from '@app-interfaces/foursquare/venue';
-import { AuthService }         from '@app-services/auth/auth.service';
-import { FourSquareService }   from '@app-services/four-square/four-square.service';
-import { UserService }         from '@app-services/user/user.service';
-import { User }                from '@app-interfaces/coffee-user';
+import { ModalController, Platform } from '@ionic/angular';
+
+import { Venue, VenueDetails }       from '@app-interfaces/foursquare/venue';
+import { AuthService }               from '@app-services/auth/auth.service';
+import { FourSquareService }         from '@app-services/four-square/four-square.service';
+import { UserService }               from '@app-services/user/user.service';
+import { CoffeeUser }                from '@app-interfaces/coffee-user';
+import { NavParams }                 from '@ionic/angular';
 
 
 
@@ -19,11 +22,11 @@ import { User }                from '@app-interfaces/coffee-user';
 })
 export class VenueDetailsCardComponent implements OnInit {
 
-  @Input('venue') 
+  //@Input('venue') 
   public venue: Venue;
 
-  @Input('user')
-  public user: User;
+  //@Input('user')
+  public user: CoffeeUser;
 
   @Output('toggleFavorite') 
   public toggleFavorite: EventEmitter<Venue> = new EventEmitter<Venue>();
@@ -31,13 +34,21 @@ export class VenueDetailsCardComponent implements OnInit {
   //details: VenueDetails = null;
   public venueCatIcon: string = '';
 
-  constructor(private auth: AuthService,private foursquare: FourSquareService, private users: UserService) {
+  constructor(private navParams: NavParams, 
+              private modalCtrl: ModalController,
+              public platform: Platform,
+              private auth: AuthService,
+              private foursquare: FourSquareService, 
+              private users: UserService) {
     
   }
 
   async ngOnInit(): Promise<void> {
-    console.log('info window', this.user);
-    console.log(this.user ? this.user.uid : 'null');
+    this.auth.user$.subscribe((u) => this.user = u)
+    this.venue = this.navParams.get('venue');
+    console.log(this.venue.location.formattedAddress.length)
+    //console.log('info window', this.user);
+    //console.log(this.user ? this.user.uid : 'null');
     //this.user$ = this.auth.user$;
     console.log('venue categories: ', this.venue.categories.length);
     const firstCat = this.venue.categories[0];
@@ -53,5 +64,11 @@ export class VenueDetailsCardComponent implements OnInit {
     catch(e) { console.log('component error: ', e) }*/
   }
 
+  async dismiss() {
+    await this.modalCtrl.dismiss();
+  }
+
   public addFavorite(): void { this.toggleFavorite.emit(this.venue) }
+
+  get isSmall(): boolean { return this.platform.width() > 692 }
 }