+++ src/app/pages/home/home.page.ts
@@ -5,7 +5,8 @@ import { Platform }                         from '@ionic/angular';
 import { Observable }                       from 'rxjs';
 
 import { AuthService }                      from '@app-services/auth/auth.service';
-import { User }                             from '@app-interfaces/coffee-user';
+import { RssFeedService }                   from '@app-services/rss-feed/rss-feed.service';
+import { CoffeeUser }                       from '@app-interfaces/coffee-user';
 
 
 @Component({
@@ -15,12 +16,20 @@ import { User }                             from '@app-interfaces/coffee-user';
 })
 export class HomePage implements OnInit {
 
-  user$: Observable<User>;
+  user$: Observable<CoffeeUser>;
+  sprudgeUrl: string = 'http://sprudge.com/feed';
 
-  constructor(private auth: AuthService, private platform: Platform) {}
+  newsList: Array<any>;
+
+  constructor(private auth: AuthService, private platform: Platform, private rssFeed: RssFeedService,) {}
 
   public async ngOnInit(): Promise<void> { 
-    this.user$ = this.auth.user$;
+    try {
+      this.user$ = this.auth.user$;
+      await this.getFeed(this.sprudgeUrl);
+      console.log(this.newsList.length);
+    }
+    catch(e) { console.log(e) }
   }
 
   public async googleLogin(): Promise<void> {
@@ -31,6 +40,14 @@ export class HomePage implements OnInit {
       }
     }
     catch(e) { console.log('googleLogin() error: ', e) }
+  }
+
+  public async getFeed(url: string) {
+    try {
+      const feed = await this.rssFeed.getFeed(url, 25);
+      this.newsList = feed.items;
+    }
+    catch(e) { console.log(e) }
   }
 
 }